Refactoring file service to new lib
