+++ src/libs/dps-files/src/test/java/ca/bc/gov/open/pssg/rsbc/dps/files/FileServiceImplTest.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.open.pssg.rsbc.dps.notification;
+package ca.bc.gov.open.pssg.rsbc.dps.files;
 
 import ca.bc.gov.open.pssg.rsbc.dps.sftp.starter.SftpService;
 import org.junit.jupiter.api.Assertions;
