+++ src/libs/dps-notification/pom.xml
@@ -17,26 +17,6 @@
 
     <dependencies>
 
-        <dependency>
-            <groupId>ca.bc.gov.open.pssg.rsbc</groupId>
-            <artifactId>dps-sftp-starter</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-
-        <!-- https://mvnrepository.com/artifact/org.junit.jupiter/junit-jupiter-api -->
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-api</artifactId>
-            <version>5.5.1</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-            <version>2.27.0</version>
-            <scope>test</scope>
-        </dependency>
-
     </dependencies>
     
 </project>
