+++ src/pom.xml
@@ -7,6 +7,7 @@
         <module>libs/figaro-ords-client</module>
         <module>libs/dps-sftp-starter</module>
         <module>libs/dps-messaging-starter</module>
+        <module>libs/dps-files</module>
     </modules>
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -78,6 +79,7 @@
             <id>all</id>
             <modules>
                 <module>libs/dps-notification</module>
+                <module>libs/dps-files</module>
                 <module>libs/figaro-ords-client</module>
                 <module>libs/vips-ords-client</module>
                 <module>libs/dps-sftp-starter</module>
@@ -120,6 +122,7 @@
             <id>spd-notification-worker</id>
             <modules>
                 <module>libs/dps-notification</module>
+                <module>libs/dps-files</module>
                 <module>libs/dps-sftp-starter</module>
                 <module>libs/dps-messaging-starter</module>
                 <module>libs/figaro-ords-client</module>
@@ -130,6 +133,7 @@
             <id>vips-notification-worker</id>
             <modules>
                 <module>libs/dps-notification</module>
+                <module>libs/dps-files</module>
                 <module>libs/dps-sftp-starter</module>
                 <module>libs/dps-messaging-starter</module>
                 <module>libs/vips-ords-client</module>
