+++ src/vips-notification-worker/pom.xml
@@ -93,6 +93,12 @@
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
 
+		<dependency>
+			<groupId>ca.bc.gov.open.pssg.rsbc</groupId>
+			<artifactId>dps-files</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
+
 	</dependencies>
 
 	<build>
