+++ src/vips-notification-worker/src/main/java/ca/bc/gov/open/pssg/rsbc/vips/notification/worker/configuration/FileConfiguration.java
@@ -1,7 +1,7 @@
 package ca.bc.gov.open.pssg.rsbc.vips.notification.worker.configuration;
 
-import ca.bc.gov.open.pssg.rsbc.dps.notification.FileService;
-import ca.bc.gov.open.pssg.rsbc.dps.notification.FileServiceImpl;
+import ca.bc.gov.open.pssg.rsbc.dps.files.FileService;
+import ca.bc.gov.open.pssg.rsbc.dps.files.FileServiceImpl;
 import ca.bc.gov.open.pssg.rsbc.dps.sftp.starter.SftpService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
