+++ src/vips-notification-worker/src/test/java/ca/bc/gov/open/pssg/rsbc/vips/notification/worker/OutputNotificationConsumerTest.java
@@ -1,8 +1,8 @@
 package ca.bc.gov.open.pssg.rsbc.vips.notification.worker;
 
-import ca.bc.gov.open.pssg.rsbc.dps.notification.FileInfo;
-import ca.bc.gov.open.pssg.rsbc.dps.notification.FileService;
-import ca.bc.gov.open.pssg.rsbc.dps.notification.OutputNotificationMessage;
+import ca.bc.gov.open.pssg.rsbc.dps.files.notification.FileInfo;
+import ca.bc.gov.open.pssg.rsbc.dps.files.notification.FileService;
+import ca.bc.gov.open.pssg.rsbc.dps.files.notification.OutputNotificationMessage;
 import ca.bc.gov.open.pssg.rsbc.dps.sftp.starter.SftpProperties;
 import ca.bc.gov.open.pssg.rsbc.dps.sftp.starter.SftpService;
 import ca.bc.gov.open.pssg.rsbc.dps.vips.notification.worker.generated.models.Data;
